requirements.txt

fastapi
uvicorn
httpx
pydantic
openai-whisper
numpy
torch
transformers
